git-svn-id: https://svn.apache.org/repos/asf/cxf/tags/cxf-2.5.2@1233587 13f79535-47bb-0310-9956-ffa450edef68

+++ rt/bindings/corba/src/main/java/org/apache/cxf/binding/corba/runtime/CorbaObjectWriter.java
@@ -183,7 +183,7 @@ public class CorbaObjectWriter {
 
     public void writeLong(Integer l) throws CorbaBindingException {
         if (l == null) {
-            stream.write_long(0);
+            stream.write_long((int)0);
         } else {
             stream.write_long(l.intValue());
         }
@@ -191,7 +191,7 @@ public class CorbaObjectWriter {
 
     public void writeULong(Long l) throws CorbaBindingException {
         if (l == null) {
-            stream.write_ulong(0);
+            stream.write_ulong((int)0);
         } else {
             stream.write_ulong(l.intValue());
         }
@@ -199,7 +199,7 @@ public class CorbaObjectWriter {
 
     public void writeLongLong(Long l) throws CorbaBindingException {
         if (l == null) {
-            stream.write_longlong(0);
+            stream.write_longlong((long)0);
         } else {
             stream.write_longlong(l.longValue());
         }
@@ -207,7 +207,7 @@ public class CorbaObjectWriter {
 
     public void writeULongLong(java.math.BigInteger l) throws CorbaBindingException {
         if (l == null) {
-            stream.write_ulonglong(0);
+            stream.write_ulonglong((long)0);
         } else {
             stream.write_ulonglong(l.longValue());
         }
@@ -223,7 +223,7 @@ public class CorbaObjectWriter {
 
     public void writeDouble(Double d) throws CorbaBindingException {
         if (d == null) {
-            stream.write_double(0.0);
+            stream.write_double((double)0.0);
         } else {
             stream.write_double(d.doubleValue());
         }
