+++ systests/ws-specs/src/test/java/org/apache/cxf/systest/ws/policy/AddressingAnonymousPolicyTest.java
@@ -137,8 +137,8 @@ public class AddressingAnonymousPolicyTest extends AbstractBusClientServerTestBa
             greeter.pingMe();
             fail("Expected PingMeFault not thrown.");
         } catch (PingMeFault ex) {
-            assertEquals(2, ex.getFaultInfo().getMajor());
-            assertEquals(1, ex.getFaultInfo().getMinor());
+            assertEquals(2, (int)ex.getFaultInfo().getMajor());
+            assertEquals(1, (int)ex.getFaultInfo().getMinor());
         } 
     }
 }
