+++ systests/ws-specs/src/test/java/org/apache/cxf/systest/ws/policy/AddressingOptionalPolicyTest.java
@@ -144,8 +144,8 @@ public class AddressingOptionalPolicyTest extends AbstractBusClientServerTestBas
             greeter.pingMe();
             fail("Expected PingMeFault not thrown.");
         } catch (PingMeFault ex) {
-            assertEquals(2, ex.getFaultInfo().getMajor());
-            assertEquals(1, ex.getFaultInfo().getMinor());
+            assertEquals(2, (int)ex.getFaultInfo().getMajor());
+            assertEquals(1, (int)ex.getFaultInfo().getMinor());
         } 
         
         MessageFlow mf = new MessageFlow(out.getOutboundMessages(), in.getInboundMessages());
@@ -194,8 +194,8 @@ public class AddressingOptionalPolicyTest extends AbstractBusClientServerTestBas
             greeter.pingMe();
             fail("Expected PingMeFault not thrown.");
         } catch (PingMeFault ex) {
-            assertEquals(2, ex.getFaultInfo().getMajor());
-            assertEquals(1, ex.getFaultInfo().getMinor());
+            assertEquals(2, (int)ex.getFaultInfo().getMajor());
+            assertEquals(1, (int)ex.getFaultInfo().getMinor());
         } 
         
         MessageFlow mf = new MessageFlow(out.getOutboundMessages(), in.getInboundMessages());
