+++ systests/ws-specs/src/test/java/org/apache/cxf/systest/ws/policy/RMPolicyTest.java
@@ -155,8 +155,8 @@ public class RMPolicyTest extends AbstractBusClientServerTestBase {
             greeter.pingMe();
             fail("Expected PingMeFault not thrown.");
         } catch (PingMeFault ex) {
-            assertEquals(2, ex.getFaultInfo().getMajor());
-            assertEquals(1, ex.getFaultInfo().getMinor());
+            assertEquals(2, (int)ex.getFaultInfo().getMajor());
+            assertEquals(1, (int)ex.getFaultInfo().getMinor());
         } 
 
         MessageRecorder mr = new MessageRecorder(outRecorder, inRecorder);
